requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
twilio==8.13.0
sqlalchemy==2.0.20
flask-sqlalchemy==3.0.5
gunicorn==20.1.0
psycopg2==2.9.9